Project Skeleton
This creates the structure of the project. I am starting out already
namespacing the routes in order to consolidate the logic. Since there
are different checks that need to be made, I think that namespacing them
will be a good idea. This will allow for better structure of the checks
as they grow.