Route to check printer api response
This adds a test for a new route to check the repsonse code of a call to
the printer api endpoint for a valid fulfiller.

This is just the skeleton for the creation of the route. There is no
logic in the route yet but this just sets up the test and Rakefile to
allow running of the tests.